requirements.txt
altair==5.5.0
attrs==25.3.0
bcrypt==4.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==6.2.0
captcha==0.7.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
extra-streamlit-components==0.1.81
gitdb==4.0.12
GitPython==3.1.45
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==2.3.0
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pip==25.2
protobuf==6.31.1
pyarrow==21.0.0
pycparser==2.22
pydeck==0.9.1
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-learn==1.6.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
streamlit-authenticator==0.4.2
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
zipp==3.23.0
zstandard==0.24.0
